refactor: split model portfolio ingestion dtos
Evidence:

- python -m pytest tests/unit/services/ingestion_service/test_reference_data_dto.py -q: 35 passed

- python -m ruff check scoped DTO/test files: passed

- python -m ruff format --check scoped DTO/test files: passed

- python -m mypy src/services/ingestion_service/app/DTOs/reference_data_model_portfolio_dto.py: passed

- radon mi: reference_data_dto.py improved from B (9.31) to B (12.49); reference_data_model_portfolio_dto.py A (41.43)
